requirements.txt
# Core runtime
requests>=2.32
jsonschema>=4.26
networkx>=3.5
pyyaml>=6.0
pydantic>=2.8

# Interactive UI (help_e.ui.server)
fastapi>=0.115
uvicorn>=0.30

# Optional extras (uncomment if you want them)
# wandb>=0.25                # experiment tracking
# sentence-transformers>=5.0 # used by data/seed_profiles.py for offline profile generation
# nltk>=3.9                  # auxiliary text utilities
